Version adaptable à windows : décommenter tous les includes stdafx, decommenter le define windows dans utile.h, gérer l'arborescence comme définie (le dossier cryptopp564 à côté du projet du programme principal, les dossiers de tests aux bons endroits...) ou modifier les paths dans le programme. Il faut aussi compiler la version 5.6.4 de la bibliothèque cryptocpp disponible ici : https://www.cryptopp.com/downloads.html, selon les instructions de ce site : http://programmingknowledgeblog.blogspot.com/2013/04/compiling-and-integrating-crypto-into.html. Le projet doit être compilé en mode release. Contrairement à ce que le site précédent indique, l'option runtime library ne doit pas être : Multi-threaded DLL mais Multi-threaded tout cours.

+++ Mercenne twister/Matrice.h
@@ -1,6 +1,8 @@
 #ifndef MATRICE_H
 #define MATRICE_H
 
+//#include "../cryptopp810/test111/stdafx.h"
+#include "utile.h"
 #include <string>
 
 using namespace std;
@@ -42,8 +44,8 @@ class Matrice
         static Matrice decalageGauche(short tx, short decalage);
 
     //private:
-        short tx;           //Nombre de colonnes
-        short ty;           //Nombre de lignes
+        int tx;             //Nombre de colonnes
+        int ty;             //Nombre de lignes
         uint8_t **elements;
 };
